+++ CHANGELOG.md
@@ -5,12 +5,14 @@
 ### Added
 
 ### Changed
+- show jump down button earlier when scrolling up
 - make font-style of info messages consistent with iOS and Android #3034
 
 ### Fixed
 - fix webxdc: allow `self` and `blob:` in `connect-src` in CSP
 - indentation in update device message
 - fix "no messages" blinking up for a second when loading a chat
+- hide unread count on jump down button if you are at the bottom, fixes #3033
 
 <a id="1_34_0"></a>